Moving the loading spinner and the dropzone components to the components folder
